requirements.txt
azure-functions
requests>=2.31.0
openai>=1.0.0
azure-cosmos>=4.5.0